Feat: Modify image upload code
